+++ src-tauri/src/app.rs
@@ -1,15 +1,14 @@
 use log::info;
-use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
-use std::ffi::OsStr;
-use std::fs;
+
+use std::collections::HashMap;
 use std::path::PathBuf;
 use std::sync::{Arc, Mutex};
+use std::thread;
 use tauri::{State, Window};
 
-use crate::dirstat::{DiskItem, DiskItemMetadata, FileInfo};
+use crate::dirstat::{DiskItem, FileInfo};
 use crate::metrics::Metrics;
 use crate::models::*;
-use rayon::prelude::*;
 
 pub struct AppState(Arc<Mutex<App>>);
 
@@ -22,9 +21,17 @@ impl AppState {
 #[derive(Default)]
 pub struct App {
     pub metrics: Metrics,
+    pub disk_analysis: HashMap<String, DiskItem>,
 }
 
 impl AppState {
+    pub fn set_disk_analysis(&self, name: String, disk_analysis: DiskItem) {
+        self.0
+            .lock()
+            .unwrap()
+            .disk_analysis
+            .insert(name, disk_analysis);
+    }
     pub fn emit_sysinfo(&self, window: &Window) {
         let sys_info = self.0.lock().unwrap().metrics.get_system_information();
         window.emit("emit_sysinfo", &sys_info).unwrap();
@@ -132,12 +139,13 @@ pub fn delete_folder(path: String) {
     }
 }
 
-// Result<Vec<FileEntry>, String>
-
 #[tauri::command]
-pub async fn deep_scan(path: String) -> Result<DiskItem, String> {
+// Multithreaded fast version, uses high cpu/memory
+pub fn disk_analysis(state: tauri::State<AppState>, path: String) -> Result<DiskItem, String> {
     let time = std::time::Instant::now();
     dbg!("Scanning folder:", &path);
+    dbg!("Triggering Disk Analysis");
+
     let path_buf = PathBuf::from(&path);
     let file_info = FileInfo::from_path(&path_buf, true).map_err(|e| e.to_string())?;
 
@@ -148,47 +156,34 @@ pub async fn deep_scan(path: String) -> Result<DiskItem, String> {
         _ => return Err("Not a directory".into()),
     };
 
+    state.set_disk_analysis(path, analysed.clone());
+
     dbg!("Scanning complete");
     dbg!("Time taken:", time.elapsed().as_secs_f32());
 
     Ok(analysed)
 }
 
-// #[tauri::command]  355.9603 seconds
-// pub async fn deep_scan(path: String) -> Result<Vec<DiskItem>, String> {
-//     let time = std::time::Instant::now();
-//     dbg!("Scanning folder:", &path);
-//     let path_buf = PathBuf::from(&path);
-//     let file_info = FileInfo::from_path(&path_buf, true).map_err(|e| e.to_string())?;
+#[tauri::command]
+// Multithreaded fast version, uses high cpu/memory
+pub async fn disk_analysis_flattened<'a>(
+    state: tauri::State<'a, AppState>,
+    path: String,
+) -> Result<Vec<DiskItem>, String> {
+    let time = std::time::Instant::now();
+    dbg!("Starting Disk Analysis flattened", time);
+    let analysis_map = &state.0.lock().map_err(|e| e.to_string())?.disk_analysis;
+    let analysis = analysis_map.get(&path).cloned();
 
-//     let analysed = match file_info {
-//         FileInfo::Directory { volume_id } => {
-//             let sub_entries = fs::read_dir(path_buf)
-//                 .map_err(|e| e.to_string())?
-//                 .filter_map(Result::ok)
-//                 .collect::<Vec<_>>();
-
-//             let mut sub_items = sub_entries
-//                 .par_iter() // Use rayon's parallel iterator
-//                 .filter_map(|entry| DiskItem::from_analyze(&entry.path(), true, volume_id).ok())
-//                 .collect::<Vec<_>>();
-
-//             sub_items.sort_unstable_by(|a, b| a.metadata.size.cmp(&b.metadata.size).reverse());
-
-//             sub_items
-//         }
-//         FileInfo::File { size, .. } => vec![DiskItem {
-//             name: path_buf
-//                 .file_name()
-//                 .unwrap_or(OsStr::new("."))
-//                 .to_string_lossy()
-//                 .to_string(),
-//             metadata: DiskItemMetadata { size },
-//             children: None,
-//         }],
-//     };
-
-//     dbg!("Scanning complete");
-//     dbg!("Time taken:", time.elapsed().as_secs_f32());
-//     Ok(analysed)
-// }
+    match analysis {
+        Some(analysis) => {
+            dbg!("Analysis found");
+            dbg!("Time taken:", time.elapsed().as_secs_f32());
+            Ok(analysis.flatten(500))
+        }
+        None => {
+            dbg!("Analysis not found");
+            Err("Analysis not found".into())
+        }
+    }
+}