+++ src-tauri/src/main.rs
@@ -21,6 +21,7 @@ mod app;
 mod dirstat;
 mod metrics;
 mod models;
+pub mod tree;
 mod utils;
 
 use app::AppState;
@@ -108,8 +109,9 @@ fn build_and_run_app(app: AppState) {
         .invoke_handler(tauri::generate_handler![
             app::kill_process,
             app::show_folder,
-            app::deep_scan,
-            app::delete_folder
+            app::delete_folder,
+            app::disk_analysis,
+            app::disk_analysis_flattened,
         ])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");