+++ src/features/metrics/components/disks/disk.directory-treeview.tsx
@@ -1,10 +1,10 @@
 import "@/features/metrics/styles/disk-treeview.css";
 
-import TreeView, { INode } from "react-accessible-treeview";
-import { IFlatMetadata } from "react-accessible-treeview/dist/TreeView/utils";
+import { NodeRendererProps, Tree } from "react-arborist";
 
 import formatBytes from "@/features/metrics/utils/format-bytes";
-import { Group, ScrollArea, Text } from "@mantine/core";
+import { DiskItem } from "@/lib";
+import { Group, Text } from "@mantine/core";
 import { IconFile, IconFolderCancel, IconFolderOpen } from "@tabler/icons-react";
 
 const iconStyle = { paddingRight: "5px", verticalAlign: "middle" };
@@ -22,35 +22,44 @@ const FileIcon = ({ fileName }: { fileName: string }) => {
   }
 };
 
+const Node = ({ node, style, dragHandle, tree, preview }: NodeRendererProps<DiskItem>) => {
+  return (
+    <div style={style} className="node" ref={dragHandle} onClick={() => node.isInternal && node.toggle()}>
+      <Group position="apart" noWrap>
+        <div>
+          {node.isLeaf ? <FileIcon fileName={node.data.name} /> : <FolderIcon isOpen={node.isOpen} />}
+          {node.data.name}
+        </div>
+
+        <Text size="xs" color="dimmed">
+          {formatBytes(node.data.size as number)}
+        </Text>
+      </Group>
+    </div>
+  );
+};
+
 interface DiskDirectoryTreeViewProps {
-  data: INode<IFlatMetadata>[];
+  data: any;
 }
 const DiskDirectoryTreeView: React.FC<DiskDirectoryTreeViewProps> = (props) => {
   const { data } = props;
 
   return (
-    <ScrollArea h="90%">
-      <TreeView
-        aria-label="directory Tree"
-        togglableSelect
-        clickAction="EXCLUSIVE_SELECT"
-        multiSelect
-        data={data}
-        nodeRenderer={({ element, isBranch, isExpanded, getNodeProps, level, handleExpand }) => {
-          return (
-            <Group position="apart">
-              <div {...getNodeProps({ onClick: handleExpand })} style={{ paddingLeft: 20 * (level - 1) }}>
-                {isBranch ? <FolderIcon isOpen={isExpanded} /> : <FileIcon fileName={element.name} />}
-                {element.name}
-              </div>
-              <Text size="xs" color="dimmed">
-                {formatBytes(element.metadata?.size as number)}
-              </Text>
-            </Group>
-          );
-        }}
-      />
-    </ScrollArea>
+    <Tree
+      data={data}
+      openByDefault={false}
+      width={"100%"}
+      height={290}
+      indent={24}
+      rowHeight={36}
+      overscanCount={1}
+      paddingTop={30}
+      paddingBottom={10}
+      padding={25}
+    >
+      {Node}
+    </Tree>
   );
 };