+++ src/features/metrics/components/disks/disk.stats-ring.tsx
@@ -1,24 +1,22 @@
-import React from 'react';
+import React from "react";
 
-import StatsRing from '@/components/stats-ring';
-import formatOverallStats from '@/features/metrics/utils/format-overall-stats';
-import useServerEventsContext from '@/hooks/useServerEventsContext';
-import { useMantineTheme } from '@mantine/core';
-import { IconCpu2, IconFileAnalytics, IconFolders } from '@tabler/icons-react';
+import StatsRing from "@/components/stats-ring";
+import useDisksStore from "@/features/metrics/stores/disk.store";
+import formatOverallStats from "@/features/metrics/utils/format-overall-stats";
+import { useMantineTheme } from "@mantine/core";
+import { IconFolders } from "@tabler/icons-react";
 
 const DiskStatsRing: React.FC = ({}) => {
-  const { disks } = useServerEventsContext();
+  const [disk] = useDisksStore.use.disks();
   const { other } = useMantineTheme();
 
-  const disk = disks[0];
+  const available = disk.total || 0;
+  const used = disk.used || 0;
 
-  const available = disk?.data?.at(-1)?.total || 0;
-  const used = disk?.data?.at(-1)?.used || 0;
-
-  // TODO: MOVE THIS TO RUST BACKEND
+  // TODO: Use USED PERCENTAGE property
   const progress = (used / available) * 100;
   const stats = React.useMemo(() => formatOverallStats(used, available), [used, available]);
-  const label = `Disk ${disk?.data?.at(-1)?.name}`;
+  const label = `Disk ${disk.name}`;
 
   return (
     <StatsRing color={other.charts.statsRing.disk} Icon={IconFolders} stats={stats} label={label} progress={progress} />