+++ src/providers/index.tsx
@@ -1,9 +1,8 @@
-import { Geiger } from 'react-geiger';
+import { Geiger } from "react-geiger";
 
-import AnalyticsProvider from '@/providers/analytics.provider';
-import ServerEventsProvider from '@/providers/server-events.provider';
-import ThemeProvider from '@/providers/theme.provider';
-import { Notifications } from '@mantine/notifications';
+import AnalyticsProvider from "@/providers/analytics.provider";
+import ThemeProvider from "@/providers/theme.provider";
+import { Notifications } from "@mantine/notifications";
 
 interface AppProvider {
   children: React.ReactNode;
@@ -14,10 +13,8 @@ const AppProvider: React.FC<AppProvider> = ({ children }) => {
     <Geiger renderTimeThreshold={50}>
       <AnalyticsProvider>
         <ThemeProvider>
-          <ServerEventsProvider>
-            <Notifications />
-            {children}
-          </ServerEventsProvider>
+          <Notifications />
+          {children}
         </ThemeProvider>
       </AnalyticsProvider>
     </Geiger>