+++ vite.config.ts
@@ -1,6 +1,7 @@
-import { defineConfig, optimizeDeps } from "vite";
-import react from "@vitejs/plugin-react";
 import * as path from "path";
+import { defineConfig, optimizeDeps } from "vite";
+
+import react from "@vitejs/plugin-react-swc";
 
 // https://vitejs.dev/config/
 export default defineConfig({